Zad 12, Zad 13, Zad 14

+++ src/com/company/Motorcycle.java
@@ -1,5 +1,5 @@
 package com.company;
 
 public class Motorcycle extends Vehicle {
-    Double yearOfProduction;
+    Integer yearOfProduction;
 }
